Paging: Transform Pager class to createPager function
